requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
datasets==4.5.0
dill==0.4.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.3
idna==3.11
joblib==1.5.3
llvmlite==0.46.0
multidict==6.7.0
multiprocess==0.70.18
numba==0.63.1
numpy==2.3.5
packaging==26.0
pandas==3.0.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyarrow==23.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
shap==0.50.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
threadpoolctl==3.6.0
tqdm==4.67.1
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
xgboost==3.1.3
xxhash==3.6.0
yarl==1.22.0
